feat: rename namespaces.py -> cleanup.py
